+++ Editor/Serialization/PropertyBag.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
+using static NewGraph.GraphSettingsSingleton;
 
 namespace NewGraph {
     /// <summary>
@@ -33,7 +34,7 @@ namespace NewGraph {
         private PropertyBag(NodeAttribute nodeAttribute, Type nodeType, SerializedProperty nodeProperty) {
             this.nodeType = nodeType;
 
-            inputPort = new PortInfo(nodeProperty.propertyPath, nodeType, new PortBaseAttribute(nodeAttribute.inputPortName, nodeAttribute.inputPortCapacity, PortDirection.Input), GraphSettings.Instance.defaultInputName);
+            inputPort = new PortInfo(nodeProperty.propertyPath, nodeType, new PortBaseAttribute(nodeAttribute.inputPortName, nodeAttribute.inputPortCapacity, PortDirection.Input), Settings.defaultInputName);
             InitializeAttributebehaviors();
             RetrieveAll(nodeProperty);
         }