+++ Editor/Views/EdgeView.cs
@@ -1,16 +1,17 @@
 using GraphViewBase;
 using UnityEngine;
+using static NewGraph.GraphSettingsSingleton;
 
 namespace NewGraph {
     public class EdgeView : Edge {
         public EdgeView() : base() {
             EdgeWidth = EdgeWidthUnselected;
         }
-        public override int EdgeWidthSelected => GraphSettings.Instance.edgeWidthSelected;
-        public override int EdgeWidthUnselected => GraphSettings.Instance.edgeWidthUnselected;
-        public override Color ColorSelected => GraphSettings.Instance.colorSelected;
+        public override int EdgeWidthSelected => Settings.edgeWidthSelected;
+        public override int EdgeWidthUnselected => Settings.edgeWidthUnselected;
+        public override Color ColorSelected => Settings.colorSelected;
         public override Color ColorUnselected => currentUnselectedColor;
 
-        public Color currentUnselectedColor = GraphSettings.Instance.colorUnselected;
+        public Color currentUnselectedColor = Settings.colorUnselected;
     }
 }